+++ OUDS/Core/Themes/OrangeInverse/Sources/Providers/ComponentTokens/OrangeInverseThemeQuantityInputComponentTokensProvider.swift
@@ -34,7 +34,7 @@ final class OrangeInverseThemeQuantityInputComponentTokensProvider: AllQuantityI
     private nonisolated(unsafe) static var instanceCount: Int = 0
     #endif
 
-    /// Defines a provider of component tokens dedicated to `OUDSTagInput`
+    /// Defines a provider of component tokens dedicated to `OUDSQuantityInput`
     /// - Parameters:
     ///    - sizes: Provider for borders semantic tokens. If nil, a default one will be used (``OrangeInverseThemeSizeSemanticTokensProvider``)
     ///    - spaces: Provider for dimension semantic tokens. If nil, a default one will be used (``OrangeInverseThemeSpaceSemanticTokensProvider``)
@@ -58,7 +58,7 @@ final class OrangeInverseThemeQuantityInputComponentTokensProvider: AllQuantityI
 
     // ଘ( ･ω･)_/ﾟ･:*:･｡☆
     // Note: So as to help the integration of generated code produced by the tokenator
-    // the implemention of TagInputComponentTokens is not here but in Core/Themes/OrangeInverse/Values/ComponentTokens/OrangeInverseTheme+QuantityInputComponentTokens.swift
+    // the implemention of QuantityInputComponentTokens is not here but in Core/Themes/OrangeInverse/Values/ComponentTokens/OrangeInverseTheme+QuantityInputComponentTokens.swift
     // This declaration of OrangeInverseThemeQuantityInputComponentTokensProvider is here also to allow to write documentation.
 }